Split texture streaming out of MaterialLoader
MaterialLoader.Streaming.cs now holds the stream types and the whole
chain pipeline - dispatch, reads, the upload pump, storage growth and
teardown - as a partial class, leaving MaterialLoader.cs to material
and texture loading.
